// serviceng2/app/Shared/enum.ts
export enum DBOperation {
    create = 1,
    update = 2,
    delete = 3,
    deleteimage = 4,

}

export enum MessageSeverity {
    error ='error',
    success ='success',
}

export enum PageOperation {
    mainlistpage = 1,
    sublistpage = 2,
    finalorderpage = 3,
}

export enum PageContentType {
    none=0,
    simpletext = 1,
    htmlcode = 2,
    image=3,
}